style: useLanternHit 훅 분리
